Update README.md
The project provides a RESTful API that fetches information for users who are stored in a JSON file using Node.js with the Express library. It includes all endpoints that allow listing users, searching for users by ID, profession, or name, and proper error responses in case of invalid requests. The important concepts of API development with JSON data handling will be demonstrated in this project.